+++ Libraries/01-Shared/Elysium.Core.Data/DbDataReader.cpp
@@ -1,20 +1,20 @@
 #include "DbDataReader.hpp"
 
-unsigned long Elysium::Core::Data::Common::DbDataReader::GetDepth() const
+size_t Elysium::Core::Data::Common::DbDataReader::GetDepth() const
 {
 	return _NumberOfRows;
 }
-unsigned long Elysium::Core::Data::Common::DbDataReader::GetFieldCount() const
+size_t Elysium::Core::Data::Common::DbDataReader::GetFieldCount() const
 {
 	return _FieldCount;
 }
 
-long Elysium::Core::Data::Common::DbDataReader::GetRecordsAffected() const
+size_t Elysium::Core::Data::Common::DbDataReader::GetRecordsAffected() const
 {
 	return _RecordsAffected;
 }
 
-Elysium::Core::Data::Common::DbDataReader::DbDataReader(unsigned long NumberOfRows, unsigned long FieldCount, long RecordsAffected)
+Elysium::Core::Data::Common::DbDataReader::DbDataReader(size_t NumberOfRows, size_t FieldCount, size_t RecordsAffected)
 	: IDataReader(),
 	_NumberOfRows(NumberOfRows), _FieldCount(FieldCount), _RecordsAffected(RecordsAffected)
 {
